Compromise for passing a config file to flake8
Issue #66 is something that I ran into, too, and when I looked at
the code configFileName wasn't even implemented.

I'm also not a flake8 expert, but it looks like flake8 has some
default locations that it looks in for config files, but what
I really want is the ability to add a project specific flake8
config file.

I removed the unused config file config parameter and added
a projectConfigFile setting that won't conflict with any previous
config settings that might exist (they'll remain no-ops).

There isn't a load of error checking in the code so I didn't try
too hard to bulletproof this, and it works perfectly for my needs.

If anyone else wants it, feel free to accept, modify, whatever.