Handle errors more gracefully
Previously we were throwing almost a raw error message to the user. Trap
this and generate a "linter message" stating that an invalid point was
encountered. Details of the original message are included as traces on the
message. In addition, a link is given sending the user to the new isue
page with some debugging information, and prompting for code sample.